stash smiles (set of 100), run inchi_a (set of 100)
With the execption of one entry, the other 99 of the set of 100
molecules are assigned a different InChI (primary_inchi) prior vs.
after (secondary_inchi).  At the level of SMILES, the cause is not
yet idenitified.  However, when reading Jmol's assigned SMILES,
among the models failing, nitro group and ester of phosphonic acid
are identified again as potentially problematic.